requirements.txt
flask
requests
python-dotenv
openai
langchain
faiss-cpu
torch
transformers
pillow
sqlalchemy